% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{analyzeByCondition}
\alias{analyzeByCondition}
\title{Perform TVDSM analysis on a dyad by condition}
\usage{
analyzeByCondition(f1 = "", f2 = "", codes = "", useRealTime = F,
  type = 2, cols = c("EDA"), dname = "Dyad", p1.name = "Participant 1",
  p2.name = "Participant 2", lag = 0, plotParams = T, downsample = 1,
  func = computeStateSpace, verbose = F, start = "", end = "",
  timeformat = "\%m/\%d/\%y \%H:\%M:\%S")
}
\arguments{
\item{f1}{path to csv file with data for participant 1}

\item{f2}{path to csv file with data for participant 2}

\item{codes}{path to csv file with condition timing data}

\item{useRealTime}{treat condition times listed in condition timing sheet as calendar times, e.g. "1/2/16 13:23:45"}

\item{type}{the type of TVDSM analysis to use}

\item{cols}{a vector indicating which columns to select from each file when creating a dyad}

\item{dname}{a string indicating a friendly name for the dyad (used in plots and summary data frame)}

\item{p1.name}{a string indicating a friendly name for the participant whose data was loaded at path indicated by \code{f1}}

\item{p2.name}{a string indicating a friendly name for the participant whose data was loaded at path indicated by \code{f2}}

\item{lag}{the lag to be used in the TVDSM analysis (in seconds)}

\item{plotParams}{should model coefficients be ploted}

\item{downsample}{the desired sample period in seconds after downsampling (e.g. 8Hz = 0.125, 2Hz = 0.5, 1hz= 1.0)}

\item{lag}{the lag to be used in the TVDSM analysis (in seconds)}
}
\value{
a list containing models (mdls), plot (plt), and a summary data frame (summary)
}
\description{
Perform TVDSM analysis on a dyad by condition
}
\examples{
analyzeByCondition("PersonA.csv","PersonB.csv",codes="ConditionTimes.csv")
}

