% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{analyzeDyad}
\alias{analyzeDyad}
\title{Perform TVDSM analysis on a dyad by condition}
\usage{
analyzeDyad(f1 = "", f2 = "", dyad = c(), xname = f1, yname = f2,
  norm = F, window_size = 60 * 5, window_step = window_size, start = "",
  end = "", func = computeStateSpace, na.rm = T, simulate = F,
  dname = paste(xname, yname, sep = "+"), lag = 0, noPlots = F,
  plotParams = T, pltTitle = paste(dname, "(", "Lag", "=", lag, ")"),
  measure = "EDA", type = 2, downsample = 1, x.baseline = NA,
  y.baseline = NA, verbose = F)
}
\arguments{
\item{f1}{path to csv file with data for participant 1}

\item{f2}{path to csv file with data for participant 2}

\item{dyad}{if a valid dyad object is provided (produced via \code{\link{as.dyad}}), analysis will be run on the dyad directly, and \code{f1} and \code{f2} will be ignored}

\item{xname}{a string indicating a friendly name for the participant whose data was loaded at path indicated by \code{f1}}

\item{yname}{a string indicating a friendly name for the participant whose data was loaded at path indicated by \code{f2}}

\item{norm}{should participant data be standardized prior to TVDSM analysis}

\item{window_size}{the size of the sliding window to use for TVDSM analysis, in seconds}

\item{window_step}{the increment to be used when sliding \code{window_size}, in seconds}

\item{dname}{a string indicating a friendly name for the dyad (used in plots and summary data frame)}

\item{lag}{the lag to be used in the TVDSM analysis (in seconds)}

\item{noPlots}{if true, plots will not be generated}

\item{plotParams}{should model coefficients be ploted}

\item{pltTitle}{pass a string to override plot title}

\item{measure}{a string indicating which column to select from each file when creating a dyad}

\item{type}{the type of TVDSM analysis to use}

\item{downsample}{the desired sample period in seconds after downsampling (e.g. 8Hz = 0.125, 2Hz = 0.5, 1hz= 1.0)}

\item{x.baseline}{provide a value to be treated as baseline for participant 1 in TVDSM analysis. Defaults to mean of participant 1's data.}

\item{y.baseline}{provide a value to be treated as baseline for participant 2 in TVDSM analysis. Defaults to mean of participant 2's data.}

\item{downsample}{the desired sample period in seconds after downsampling (e.g. 8Hz = 0.125, 2Hz = 0.5, 1hz= 1.0)}
}
\value{
a list containing models (mdls), plot (plt), and a summary data frame (summary)
}
\description{
Perform TVDSM analysis on a dyad by condition
}
\examples{
#Perform TVDSM analysis using a 5 minute sliding window incremented every 10 second, with a lag of 5 seconds
analyzeDyad("PersonA.csv","PersonB.csv",window_size=5*60, window_step=10,lag=5)
}

