% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate.relative}
\alias{annotate.relative}
\title{Annotate a ggplot with relative coordinates}
\usage{
annotate.relative(plt = last_plot(), geom = "text", x = 0.5, y = 0.1, ...)
}
\arguments{
\item{plt}{A ggplot object. Defaults to the last plot.}

\item{geom}{Character string specifying the type of geom to add; defaults to 'text'.}

\item{x}{Relative x-position (0 to 1). Default is 0.5.}

\item{y}{Relative y-position (0 to 1). Default is 0.1.}

\item{...}{Additional arguments passed to the annotate function.}
}
\value{
The annotation layer added to the plot.
}
\description{
This function adds an annotation (default text) at a relative position on the ggplot.
}
\examples{
annotate.relative(plt, label = 'Hello')
}
