% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{as.dyad}
\alias{as.dyad}
\title{Convert Two Participant Data into a Dyad}
\usage{
as.dyad(
  p1,
  p2,
  cols = c("EDA"),
  norm = F,
  verbose = F,
  na.interpolate = T,
  interpolation.method = na.spline
)
}
\arguments{
\item{p1}{Data frame for participant 1.}

\item{p2}{Data frame for participant 2.}

\item{cols}{Columns to include (default "EDA").}

\item{norm}{Logical, if TRUE normalize the data (default FALSE).}

\item{verbose}{Logical for verbose output (default FALSE).}

\item{na.interpolate}{Logical, if TRUE interpolate NA values (default TRUE).}

\item{interpolation.method}{Function to interpolate NAs (default na.spline).}
}
\value{
A dyadic data frame with aligned timestamps.
}
\description{
Aligns and merges two participant data frames into a dyadic data frame based on overlapping timestamps.
}
