% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{as.simulateddyad}
\alias{as.simulateddyad}
\title{Create Simulated Dyad Data}
\usage{
as.simulateddyad(p1, p2, cols = c("EDA"), norm = F, verbose = F)
}
\arguments{
\item{p1}{Data frame for participant 1.}

\item{p2}{Data frame for participant 2.}

\item{cols}{Columns to include (default "EDA").}

\item{norm}{Logical indicating whether to normalize data (default FALSE).}

\item{verbose}{Logical for verbose output (default FALSE).}
}
\value{
A data frame representing the simulated dyad.
}
\description{
Simulates a dyad using two participant data frames.
}
