% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{compareSim}
\alias{compareSim}
\title{Compare Simulated Signals}
\usage{
compareSim(
  x.signal = sampleSignal(),
  fs = 0.1,
  coReg.coef = 0.1,
  selfReg.coef = 0.05
)
}
\arguments{
\item{x.signal}{Signal for person A (default: sampleSignal()).}

\item{fs}{Sampling frequency (default: 0.1).}

\item{coReg.coef}{Co-regulation coefficient (default: 0.1).}

\item{selfReg.coef}{Self-regulation coefficient (default: 0.05).}
}
\value{
A plot grid comparing simulation outputs.
}
\description{
Compares simulated signals using partner simulation and dyadic analysis, then plots the results.
}
