% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{describeDyad}
\alias{describeDyad}
\title{Describe dyad interdependence}
\usage{
describeDyad(mdls, xname = "Participant 1", yname = "Participant 2")
}
\arguments{
\item{mdls}{A list of TVDSM model outputs.}

\item{xname}{Friendly name for participant 1.}

\item{yname}{Friendly name for participant 2.}
}
\value{
A summary of interdependence measures.
}
\description{
Computes descriptive statistics for dyad interdependence metrics based on TVDSM outputs.
}
