% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{descriptivesByCondition}
\alias{descriptivesByCondition}
\title{Descriptives By Condition}
\usage{
descriptivesByCondition(
  filename,
  codes,
  col = "EDA",
  title = filename,
  codeFunc = read.RTCodes,
  startCol = "Start.Time",
  endCol = "End.Time",
  conditionCol = "Condition",
  timeformat = "\%Y-\%m-\%d \%H:\%M:\%S"
)
}
\arguments{
\item{filename}{Path to the EDA data file.}

\item{codes}{Codes data (either file path or dataframe).}

\item{col}{Column for analysis (default "EDA").}

\item{title}{Title for the plot.}

\item{codeFunc}{Function to read codes, default is read.RTCodes.}

\item{startCol}{Start time column name, default "Start.Time".}

\item{endCol}{End time column name, default "End.Time".}

\item{conditionCol}{Condition column name, default "Condition".}

\item{timeformat}{Time format string.}
}
\value{
Data frame of descriptive statistics.
}
\description{
Descriptives By Condition
}
