% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{distance}
\alias{distance}
\title{Calculate Euclidean Distance}
\usage{
distance(x, y)
}
\arguments{
\item{x}{Numeric value.}

\item{y}{Numeric value.}
}
\value{
Euclidean distance computed as sqrt(x*x + y*y).
}
\description{
Calculate Euclidean Distance
}
