% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{downsample.eda}
\alias{downsample.eda}
\title{Downsample EDA Data}
\usage{
downsample.eda(d, targetRate = 1)
}
\arguments{
\item{d}{Data frame with EDA data.}

\item{targetRate}{Desired sampling rate (default 1.0).}
}
\value{
A downsampled data frame.
}
\description{
Downsamples EDA data to a target sampling rate.
}
