% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{dsmodel.multilag}
\alias{dsmodel.multilag}
\title{Fit Multi-lag Dyadic Model}
\usage{
dsmodel.multilag(x, y, step = 0.25, p.value = 0.01, type = 2)
}
\arguments{
\item{x}{Numeric vector for signal x.}

\item{y}{Numeric vector for signal y.}

\item{step}{Step size (default 0.25).}

\item{p.value}{P-value threshold (default 0.01).}

\item{type}{Model type (default 2).}
}
\value{
A list containing models, statistics, and prediction functions.
}
\description{
Fit Multi-lag Dyadic Model
}
