% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{dsmodel}
\alias{dsmodel}
\title{Fit Dyadic Model}
\usage{
dsmodel(x, y, type = 2, step = 0.25, p.value = 0.01, verbose = F)
}
\arguments{
\item{x}{Numeric vector for signal x.}

\item{y}{Numeric vector for signal y.}

\item{type}{Model type (default 2).}

\item{step}{Step size (default 0.25).}

\item{p.value}{P-value threshold (default 0.01).}

\item{verbose}{Logical indicating verbose output (default FALSE).}
}
\value{
A list containing the fitted models, R-squared values, prediction functions, and model equations.
}
\description{
Fit Dyadic Model
}
