% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{dsmodel.sem}
\alias{dsmodel.sem}
\title{Fit Dyadic Model using SEM}
\usage{
dsmodel.sem(d, type = 2, step = 0.25, p.value = 0.01, verbose = T)
}
\arguments{
\item{d}{Data frame containing dyad data.}

\item{type}{Model type (default 2).}

\item{step}{Step size (default 0.25).}

\item{p.value}{P-value threshold (default 0.01).}

\item{verbose}{Logical indicating verbose output (default TRUE).}
}
\value{
A SEM model object.
}
\description{
Fit Dyadic Model using SEM
}
