% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{dummyCodeByCondition}
\alias{dummyCodeByCondition}
\title{Dummy Code By Condition}
\usage{
dummyCodeByCondition(
  filename,
  codes,
  outputFile = paste(filename, "_DummyCoded.csv", sep = ""),
  saveFile = F,
  codeFunc = read.RTCodes,
  startCol = "Start.Time",
  endCol = "End.Time",
  conditionCol = "Condition",
  timeformat = "\%Y-\%m-\%d \%H:\%M:\%S"
)
}
\arguments{
\item{filename}{Filename for the EDA data.}

\item{codes}{Either a path to a file with codes or a dataframe with codes.}

\item{outputFile}{Output file name (default is filename appended with "_DummyCoded.csv").}

\item{saveFile}{Logical indicating whether to save the dummy coded file.}

\item{codeFunc}{Function to read codes, default is read.RTCodes.}

\item{startCol}{Column name for start time, default "Start.Time".}

\item{endCol}{Column name for end time, default "End.Time".}

\item{conditionCol}{Column name for condition, default "Condition".}

\item{timeformat}{Time format for conversion.}
}
\value{
Data frame with dummy coded data.
}
\description{
Dummy Code By Condition
}
