% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{dyad.dsgraph}
\alias{dyad.dsgraph}
\title{Plot Dyadic Graph}
\usage{
dyad.dsgraph(
  p1,
  p2,
  norm = T,
  epoch = 1,
  step = 0.25,
  title = "State Space Plot",
  type = 2,
  measure = "EDA"
)
}
\arguments{
\item{p1}{Data for participant 1.}

\item{p2}{Data for participant 2.}

\item{norm}{Logical indicating whether to normalize the data (default TRUE).}

\item{epoch}{Epoch size (default 1).}

\item{step}{Step size (default 0.25).}

\item{title}{Plot title.}

\item{type}{Model type (default 2).}

\item{measure}{Column name for measurement (default "EDA").}
}
\value{
A plot of the model params
}
\description{
Plot Dyadic Graph
}
