% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{dyad.dsmodel}
\alias{dyad.dsmodel}
\title{Build Dyadic Model}
\usage{
dyad.dsmodel(p1, p2, type = 2, step = 0.25, measure = "EDA")
}
\arguments{
\item{p1}{Data for participant 1.}

\item{p2}{Data for participant 2.}

\item{type}{Model type (default 2).}

\item{step}{Step size (default 0.25).}

\item{measure}{Column name for measurement (default "EDA").}
}
\value{
A model object.
}
\description{
Build Dyadic Model
}
