% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{get.key}
\alias{get.key}
\title{Extract a key from a list of objects}
\usage{
get.key(lst, key, as.list = F)
}
\arguments{
\item{lst}{A list of objects.}

\item{key}{The key to extract.}

\item{as.list}{Logical; if TRUE, returns a list (default FALSE).}
}
\value{
A vector or list of values corresponding to the key from each element.
}
\description{
Retrieves the specified key from each element of a list.
}
