% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{getFS}
\alias{getFS}
\title{Get Sampling Frequency}
\usage{
getFS(d)
}
\arguments{
\item{d}{Data frame with a Timestamp column.}
}
\value{
The sampling frequency.
}
\description{
Computes the sampling frequency from the Timestamp column of a data frame.
}
