% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggCaterpillar}
\alias{ggCaterpillar}
\title{ggCaterpillar: Plot random effects as QQ-plots or caterpillar dotplots.}
\usage{
ggCaterpillar(re, QQ = TRUE, likeDotplot = TRUE)
}
\arguments{
\item{re}{A list or matrix of random effects.}

\item{QQ}{Logical flag; if TRUE a QQ-plot is generated, otherwise a caterpillar plot is used (default: TRUE).}

\item{likeDotplot}{Logical flag; if TRUE the plot is styled like a dotplot (default: TRUE).}
}
\value{
A list of ggplot objects.
}
\description{
This function creates either a normal QQ-plot or a caterpillar-style dotplot for random effects.
}
