% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggrel.x}
\alias{ggrel.x}
\title{Calculate a relative x-coordinate in a ggplot}
\usage{
ggrel.x(plt = last_plot(), x = 0.5)
}
\arguments{
\item{plt}{A ggplot object. Defaults to the last plot.}

\item{x}{A numeric value between 0 and 1 representing the relative position. Default is 0.5.}
}
\value{
A numeric value representing the x-coordinate.
}
\description{
Given a ggplot object, this function returns an x-coordinate at a relative position between the plot limits.
}
\examples{
ggrel.x()
}
