% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggrel.y}
\alias{ggrel.y}
\title{Calculate a relative y-coordinate in a ggplot}
\usage{
ggrel.y(plt = last_plot(), y = 0.5)
}
\arguments{
\item{plt}{A ggplot object. Defaults to the last plot.}

\item{y}{A numeric value between 0 and 1 representing the relative position. Default is 0.5.}
}
\value{
A numeric value representing the y-coordinate.
}
\description{
Given a ggplot object, this function returns a y-coordinate at a relative position between the plot limits.
}
\examples{
ggrel.y()
}
