% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{linearTrend}
\alias{linearTrend}
\title{Compute Linear Trend Slope}
\usage{
linearTrend(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
The slope of the linear trend.
}
\description{
Compute Linear Trend Slope
}
