% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{lmp}
\alias{lmp}
\title{Calculate p-value from an lm object}
\usage{
lmp(modelobject)
}
\arguments{
\item{modelobject}{An object of class 'lm' or 'systemfit.equation'.}
}
\value{
The p-value of the F-statistic.
}
\description{
Calculate p-value from an lm object
}
