% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{normalize}
\alias{normalize}
\title{Normalize a Numeric Vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Normalized vector where the mean is subtracted and divided by the standard deviation.
}
\description{
Normalize a Numeric Vector
}
