% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{o.center}
\alias{o.center}
\title{Center a Numeric Vector}
\usage{
o.center(d)
}
\arguments{
\item{d}{A numeric vector to center.}

\item{use.sd}{Logical indicating whether to scale by standard deviation (default FALSE).}
}
\value{
A scaled numeric vector.
}
\description{
This function centers a numeric vector by subtracting its minimum (and optionally scaling by SD).
}
