% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{o.coef}
\alias{o.coef}
\title{Extract Coefficient with Significance Check}
\usage{
o.coef(mdl, c = 2, cutoff = 0.05, useD = F)
}
\arguments{
\item{mdl}{Linear model object.}

\item{c}{Parameter index (default 2).}

\item{cutoff}{Significance cutoff (default 0.05).}

\item{useD}{Logical, if TRUE use standardized coefficient (default FALSE).}
}
\value{
Coefficient value or 0 if not significant.
}
\description{
Extract Coefficient with Significance Check
}
