% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{o.estLag}
\alias{o.estLag}
\title{Estimate Lag with Moving Window}
\usage{
o.estLag(x, window = 10)
}
\arguments{
\item{x}{Numeric vector.}

\item{window}{Window size (default 10).}
}
\value{
Estimated lag values.
}
\description{
Estimate Lag with Moving Window
}
