% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{o.Lag}
\alias{o.Lag}
\title{Compute Lag Difference}
\usage{
o.Lag(x, lag = 0, robust = T)
}
\arguments{
\item{x}{Numeric vector.}

\item{lag}{Lag value.}

\item{robust}{Logical indicating use of robust method (default TRUE).}
}
\value{
Lagged difference.
}
\description{
Compute Lag Difference
}
