% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{o.scale}
\alias{o.scale}
\title{Scale Vector to [0,1]}
\usage{
o.scale(x, use.sd = F)
}
\arguments{
\item{x}{Numeric vector.}

\item{use.sd}{Logical, if TRUE uses standard deviation for scaling (default FALSE).}
}
\value{
Scaled numeric vector.
}
\description{
Scales a numeric vector to the range [0,1].
}
