% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{o.smooth}
\alias{o.smooth}
\title{Smooth a Numeric Vector}
\usage{
o.smooth(x, window = 10)
}
\arguments{
\item{x}{Numeric vector.}

\item{window}{Window size (default 10).}
}
\value{
Smoothed vector.
}
\description{
Smooth a Numeric Vector
}
