% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{p.format}
\alias{p.format}
\title{Format p-value for display}
\usage{
p.format(p)
}
\arguments{
\item{p}{Numeric, the p-value to format.}
}
\value{
A character string representing the formatted p-value.
}
\description{
Formats a p-value into a string with appropriate formatting. If the p-value is less than 0.001, it omits the equal sign.
}
\examples{
p.format(0.0005)
}
