% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{plot.dsgraph}
\alias{plot.dsgraph}
\title{Create Plot for Dyadic State Space Graph}
\usage{
\method{plot}{dsgraph}(s, title = "State Space Plot", xlabel = "X Data", ylabel = "Y Data")
}
\arguments{
\item{s}{A dyadic statespace model object.}

\item{title}{Plot title.}

\item{xlabel}{Label for x-axis.}

\item{ylabel}{Label for y-axis.}
}
\value{
A ggplot object.
}
\description{
Create Plot for Dyadic State Space Graph
}
