% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eda}
\alias{plot.eda}
\title{Plot EDA data with optional accelerometer data}
\usage{
\method{plot}{eda}(data, title = "EDA Data", show_acc = FALSE, show_raw_acc = FALSE)
}
\arguments{
\item{data}{A data frame containing at least columns 'Timestamp' and 'EDA'. Also expected are 'X', 'Y', 'Z' if accelerometer is desired.}

\item{title}{Character string for the plot title. Default is 'EDA Data'.}

\item{show_acc}{Logical. If TRUE, an additional plot for acceleration is shown.}

\item{show_raw_acc}{Logical. If TRUE, raw accelerometer data (X, Y, Z) is plotted; otherwise, the magnitude (Movement) is computed and plotted.}
}
\value{
A ggplot object or a cowplot object if accelerometer data is added.
}
\description{
This function creates a line plot for EDA data and, optionally, an additional plot for accelerometer data.
}
\examples{
plot.eda(mydata, title = 'Subject 1 EDA', show_acc = TRUE)
}
