% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDyad}
\alias{plotDyad}
\title{Plot dyadic data using ggplot2}
\usage{
plotDyad(
  d,
  title = "",
  ylabel = "EDA",
  legend_label = "Participant",
  grayscale = FALSE,
  ...
)
}
\arguments{
\item{d}{A data frame with at least a 'Timestamp' column and two other variables to compare.}

\item{title}{Plot title.}

\item{ylabel}{Label for the y-axis. Default is 'EDA'.}

\item{legend_label}{Label for the legend. Default is 'Participant'.}

\item{grayscale}{Logical. If TRUE, the plot is rendered in grayscale.}

\item{...}{Additional parameters passed to ggplot.}
}
\value{
A ggplot object.
}
\description{
This function reshapes data for two variables (dyad) and plots them as lines with distinct colors.
}
\examples{
plotDyad(mydata, title = 'Dyad Plot')
}
