% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotlyDyad}
\alias{plotlyDyad}
\title{Plot dyadic data using Plotly}
\usage{
plotlyDyad(
  d,
  title = "",
  ylabel = "EDA",
  legend_label = "Participant",
  grayscale = FALSE,
  ...
)
}
\arguments{
\item{d}{A data frame with at least a 'Timestamp' column and two variables.}

\item{title}{Plot title.}

\item{ylabel}{Label for the y-axis. Default is 'EDA'.}

\item{legend_label}{Label for the legend. Default is 'Participant'.}

\item{grayscale}{Logical. If TRUE, applies a grayscale color scheme.}

\item{...}{Additional parameters.}
}
\value{
A Plotly plot object.
}
\description{
This function reshapes data for two variables and creates an interactive Plotly line plot.
}
\examples{
plotlyDyad(mydata, title = 'Dyad Interactive Plot')
}
