% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{plotTSDescriptives}
\alias{plotTSDescriptives}
\title{Plot Time Series Descriptive Statistics}
\usage{
plotTSDescriptives(data, title = "", vars = NULL)
}
\arguments{
\item{data}{Data frame of descriptive statistics.}

\item{title}{Plot title.}

\item{vars}{Optional variables to melt.}
}
\value{
A ggplot object.
}
\description{
Plot Time Series Descriptive Statistics
}
