% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r2.corr.mer}
\alias{r2.corr.mer}
\title{Calculate R-squared from a Mixed Effects Model}
\usage{
r2.corr.mer(m)
}
\arguments{
\item{m}{A mixed effects model (e.g., from lme4).}
}
\value{
R-squared value.
}
\description{
This function fits a simple linear model of the response against the fitted values from a mixed effects model and returns the R-squared.
}
