% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{rand.dyad}
\alias{rand.dyad}
\title{Generate Random Dyadic Data}
\usage{
rand.dyad(dur = 60 * 120, fs = 32)
}
\arguments{
\item{dur}{Duration in seconds (default 60*120).}

\item{fs}{Sampling rate (default 32).}
}
\value{
Data frame with Timestamp and two fake signals.
}
\description{
Generate Random Dyadic Data
}
