% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.actiwave}
\alias{read.actiwave}
\title{Read Actiwave Data}
\usage{
read.actiwave(file)
}
\arguments{
\item{file}{File path to the Actiwave CSV file.}
}
\value{
A data frame with Timestamp, HR, and IBI columns.
}
\description{
Reads Actiwave data from a CSV file and computes IBI.
}
