% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.biosync}
\alias{read.biosync}
\title{Read Biosync Data}
\usage{
read.biosync(file)
}
\arguments{
\item{file}{File path to the CSV file.}
}
\value{
A data frame with a corrected Timestamp column.
}
\description{
Reads biosync data from a CSV file and fixes the Timestamp format.
}
