% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.e4}
\alias{read.e4}
\title{Read E4 Data}
\usage{
read.e4(path)
}
\arguments{
\item{path}{Directory path containing E4 CSV files.}
}
\value{
A data frame with processed E4 data.
}
\description{
Reads E4 device data from a specified directory.
}
