% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.eda}
\alias{read.eda}
\title{Read EDA Data}
\usage{
read.eda(file)
}
\arguments{
\item{file}{File path to the EDA data file.}
}
\value{
A data frame with EDA data, Timestamp, and Motion computed.
}
\description{
Reads EDA (electrodermal activity) data from a CSV file.
}
