% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.edf}
\alias{read.edf}
\title{Read EDF Data}
\usage{
read.edf(file)
}
\arguments{
\item{file}{File path to the EDF file.}
}
\value{
A list of data frames, one per signal.
}
\description{
Reads EDF (European Data Format) data using edfReader.
}
