% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.empatica.acc}
\alias{read.empatica.acc}
\title{Read Empatica Accelerometer Data}
\usage{
read.empatica.acc(file)
}
\arguments{
\item{file}{File path to the ACC CSV file.}
}
\value{
A data frame with X, Y, Z, and Timestamp columns.
}
\description{
Reads accelerometer data from an Empatica ACC CSV file.
}
