% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.empatica.bvp}
\alias{read.empatica.bvp}
\title{Read Empatica BVP Data}
\usage{
read.empatica.bvp(file)
}
\arguments{
\item{file}{File path to the BVP CSV file.}
}
\value{
A data frame with BVP and Timestamp columns.
}
\description{
Reads blood volume pulse (BVP) data from an Empatica BVP CSV file.
}
