% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.empatica.eda}
\alias{read.empatica.eda}
\title{Read Empatica EDA Data}
\usage{
read.empatica.eda(file)
}
\arguments{
\item{file}{File path to the EDA CSV file.}
}
\value{
A data frame with Timestamp and EDA data.
}
\description{
Reads EDA data from an Empatica EDA CSV file.
}
