% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.empatica}
\alias{read.empatica}
\title{Read Empatica Data}
\usage{
read.empatica(path)
}
\arguments{
\item{path}{Directory path containing Empatica CSV files.}
}
\value{
A list of data frames for different signals (ACC, BVP, EDA, TEMP, IBI).
}
\description{
Reads Empatica device data from a specified directory.
}
