% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.ERCodes}
\alias{read.ERCodes}
\title{Read Engagement Codes}
\usage{
read.ERCodes(path)
}
\arguments{
\item{path}{File path to the codes file.}
}
\value{
A data frame containing engagement codes.
}
\description{
Reads engagement codes from an Excel or CSV file.
}
