% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.ibutton}
\alias{read.ibutton}
\title{Read iButton Data}
\usage{
read.ibutton(path, header = T)
}
\arguments{
\item{path}{File path to the CSV file.}

\item{header}{Logical indicating if file has header (default TRUE).}
}
\value{
A data frame with Timestamp and Temperature columns.
}
\description{
Reads data from an iButton CSV file skipping header lines.
}
