% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.reda}
\alias{read.reda}
\title{Read Reda Data}
\usage{
read.reda(file)
}
\arguments{
\item{file}{File path to the REDA data file.}
}
\value{
A data frame containing the processed REDA data.
}
\description{
Reads and processes REDA data from a CSV file.
}
