% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.RTCodes}
\alias{read.RTCodes}
\title{Read RTCodes}
\usage{
read.RTCodes(
  path = "ConditionTimes.csv",
  startCol = "Start.Time",
  endCol = "End.Time",
  fmt = "\%y-\%m-\%d \%H:\%M:\%S"
)
}
\arguments{
\item{path}{File path (default "ConditionTimes.csv").}

\item{startCol}{Column name for start time.}

\item{endCol}{Column name for end time.}

\item{fmt}{Time format (default "%y-%m-%d %H:%M:%S").}
}
\value{
A data frame with RTCodes.
}
\description{
Reads RTCodes from a CSV or Excel file.
}
