% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{RMSSD}
\alias{RMSSD}
\title{Compute Root Mean Square of Successive Differences (RMSSD)}
\usage{
RMSSD(hr.period)
}
\arguments{
\item{hr.period}{A numeric vector of heart rate period values.}
}
\value{
The RMSSD value.
}
\description{
Compute Root Mean Square of Successive Differences (RMSSD)
}
