% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{runAnova}
\alias{runAnova}
\title{Run ANOVA on model set}
\usage{
runAnova(mdls, key = "Condition")
}
\arguments{
\item{mdls}{A list of model objects.}

\item{key}{The key to extract the condition variable (default "Condition").}
}
\value{
A fitted lmer model object.
}
\description{
Runs ANOVA comparisons on delta R-squared values from a list of models using a specified condition.
}
