% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{runFalsePairings}
\alias{runFalsePairings}
\title{Run False Pairings Analysis}
\usage{
runFalsePairings(
  participantAList,
  participantBList,
  window_size = 60,
  window_step = window_size,
  downsample = 1,
  lag = 0,
  type = 3
)
}
\arguments{
\item{participantAList}{List of signals for person A.}

\item{participantBList}{List of signals for person B.}

\item{window_size}{Window size for analysis (default: 60).}

\item{window_step}{Step size between windows (default: window_size).}

\item{downsample}{Downsample factor (default: 1).}

\item{lag}{Lag to be applied (default: 0).}

\item{type}{Analysis type specifier (default: 3).}
}
\value{
A list containing results for true pairs and false pairs.
}
\description{
Runs TVDSM analysis on true and false pairings from lists of participant signals.
}
