% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sampleSignal}
\alias{sampleSignal}
\title{Sample Signal}
\usage{
sampleSignal()
}
\value{
A numeric vector representing the sample signal.
}
\description{
Reads a sample signal from an Excel file for simulation purposes.
}
