% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{save.eda}
\alias{save.eda}
\title{Save EDA Data to File}
\usage{
save.eda(data, filename)
}
\arguments{
\item{data}{Data frame containing EDA data.}

\item{filename}{Output file name.}
}
\description{
Writes EDA data to a file with a header.
}
