% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a scatter plot with a linear fit and correlation annotation}
\usage{
scatter_plot(x, y, data)
}
\arguments{
\item{x}{Character string representing the name of the x variable in the data frame.}

\item{y}{Character string representing the name of the y variable in the data frame.}

\item{data}{A data frame containing the variables.}
}
\value{
A ggplot object with the scatter plot and annotations.
}
\description{
This function generates a scatter plot for two variables, adds a linear regression line, and annotates the plot with the correlation coefficient and p-value.
}
\examples{
scatter_plot('height', 'weight', mydata)
}
