% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{simulate}
\alias{simulate}
\title{Simulate Data using Linear Models}
\usage{
simulate(xmodel, ymodel, n = 1000, start = c(1, 2))
}
\arguments{
\item{xmodel}{Linear model for x simulation.}

\item{ymodel}{Linear model for y simulation.}

\item{n}{Number of samples to simulate (default 1000).}

\item{start}{Starting values as a vector (default c(1,2)).}
}
\value{
A matrix with simulated data (n rows x 2 columns).
}
\description{
Simulate Data using Linear Models
}
