% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{statespacegraph}
\alias{statespacegraph}
\title{Plot State Space Graph}
\usage{
statespacegraph(
  xt,
  yt,
  norm = T,
  step = 0.25,
  title = "State Space Plot",
  xlabel = "X Data",
  ylabel = "Y Data",
  type = 3
)
}
\arguments{
\item{xt}{Numeric vector for x values.}

\item{yt}{Numeric vector for y values.}

\item{norm}{Logical to normalize data (default TRUE).}

\item{step}{Step size (default 0.25).}

\item{title}{Plot title.}

\item{xlabel}{Label for x-axis.}

\item{ylabel}{Label for y-axis.}

\item{type}{Model type (default 3).}
}
\value{
A ggplot object representing the state space graph.
}
\description{
Plot State Space Graph
}
