% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{topography}
\alias{topography}
\title{Plot Topography Density}
\usage{
topography(x, y, xlab = "X", ylab = "Y")
}
\arguments{
\item{x}{Numeric vector for x-axis.}

\item{y}{Numeric vector for y-axis.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}
}
\value{
A ggplot object showing 2D density plot.
}
\description{
Plot Topography Density
}
