% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{tsDescriptives}
\alias{tsDescriptives}
\title{Compute Time Series Descriptive Statistics}
\usage{
tsDescriptives(data, fs = 32)
}
\arguments{
\item{data}{Numeric vector of data.}

\item{fs}{Sample rate (default is 32).}
}
\value{
A list of descriptive statistics including mean, se, sd, var, trend, min, max, slope, slopeRatio, range and autocorrelation.
}
\description{
Compute Time Series Descriptive Statistics
}
