% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{TSDescriptivesByCondition}
\alias{TSDescriptivesByCondition}
\title{Alias for descriptivesByCondition}
\usage{
TSDescriptivesByCondition(
  filename,
  codes,
  col = "EDA",
  title = filename,
  codeFunc = read.RTCodes,
  startCol = "Start.Time",
  endCol = "End.Time",
  conditionCol = "Condition",
  timeformat = "\%Y-\%m-\%d \%H:\%M:\%S"
)
}
\description{
Alias for descriptivesByCondition
}
