% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{urlForModel}
\alias{urlForModel}
\title{Generate URL for model visualization}
\usage{
urlForModel(m)
}
\arguments{
\item{m}{A model object containing coefficients and a matrix.}
}
\value{
A string representing the URL for the model visualization.
}
\description{
Constructs a local file URL for visualizing model results based on coefficients.
}
