% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{vectorfield}
\alias{vectorfield}
\title{Create a Vector Flow Plot}
\usage{
vectorfield(x, y, xlabel = "X", ylabel = "Y", title = "Vector Flow Plot")
}
\arguments{
\item{x}{Numeric vector for x coordinates.}

\item{y}{Numeric vector for y coordinates.}

\item{xlabel}{Label for x-axis.}

\item{ylabel}{Label for y-axis.}

\item{title}{Plot title.}
}
\value{
A ggplot object displaying the vector flow plot.
}
\description{
Create a Vector Flow Plot
}
