% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{vectorflow}
\alias{vectorflow}
\title{Visualize Vector Field}
\usage{
vectorflow(x, y, f, scale = 0.1)
}
\arguments{
\item{x}{Numeric vector for x-axis.}

\item{y}{Numeric vector for y-axis.}

\item{f}{Function for vector field computation.}

\item{scale}{Scaling factor (default 0.1).}
}
\description{
Visualize Vector Field
}
