% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio.R
\name{window}
\alias{window}
\title{Apply Moving Window Operation}
\usage{
window(x, window_size, window_overlap = 0, FUN, na.rm = T)
}
\arguments{
\item{x}{Numeric vector.}

\item{window_size}{Size of each window.}

\item{window_overlap}{Overlap between windows.}

\item{FUN}{Function to apply on each window.}

\item{na.rm}{Logical indicating whether to remove NAs (default TRUE).}
}
\value{
A vector of computed values.
}
\description{
Apply Moving Window Operation
}
