% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpersonal.R
\name{analyzeDyad}
\alias{analyzeDyad}
\title{Analyze dyad data for TVDSM analysis}
\usage{
analyzeDyad(
  f1 = "",
  f2 = "",
  dyad = c(),
  xname = f1,
  yname = f2,
  norm = F,
  window_size = 60 * 5,
  window_step = window_size,
  start = "",
  end = "",
  func = fitDSModelForDyad,
  na.rm = T,
  simulate = F,
  dname = paste(xname, yname, sep = "+"),
  lag = 0,
  noPlots = F,
  plotParams = T,
  pltTitle = paste(dname, "(", "Lag", "=", lag, ")"),
  measure = "EDA",
  type = 4,
  downsample = 1,
  x.baseline = NA,
  y.baseline = NA,
  verbose = F,
  codes = "",
  useRealTime = F,
  incTSD = T,
  ...
)
}
\arguments{
\item{f1}{File path for participant 1.}

\item{f2}{File path for participant 2.}

\item{dyad}{Optional dyad object.}

\item{xname}{Friendly name for participant 1.}

\item{yname}{Friendly name for participant 2.}

\item{norm}{Logical, whether to standardize data.}

\item{window_size}{Window size in seconds.}

\item{window_step}{Step size in seconds.}

\item{start}{Start time for analysis.}

\item{end}{End time for analysis.}

\item{func}{Function for model fitting (default fitDSModelForDyad).}

\item{na.rm}{Logical.}

\item{simulate}{Logical, whether to simulate a dyad.}

\item{dname}{Friendly name for the dyad.}

\item{lag}{Lag value for analysis.}

\item{noPlots}{Logical, if TRUE, no plots are produced.}

\item{plotParams}{Logical, whether to include model parameter plots.}

\item{pltTitle}{Title for the plots.}

\item{measure}{Measurement column name.}

\item{type}{Model type.}

\item{downsample}{Downsampling factor.}

\item{x.baseline}{Baseline for participant 1.}

\item{y.baseline}{Baseline for participant 2.}

\item{verbose}{Logical, whether to be verbose.}

\item{codes}{File path for condition codes.}

\item{useRealTime}{Logical, whether to interpret condition codes in real time.}

\item{incTSD}{Logical, whether to include time series descriptives.}

\item{...}{Additional arguments.}
}
\value{
A list containing TVDSM model outputs, summary, and plots.
}
\description{
Processes two participant datasets or a dyad object, performs TVDSM analysis, generates plots, and returns model summaries.
}
